add new NetPAvailabilityState logic
Locked and Unlocked are no longer sufficient for us to render the correct states.

We add a new NetPVisibilityState which tells us if we should show the VPN item to the user, and if we do, what it's subscription state is.

I'm not entirely sure about nesting the subscription status, or having this as a property of Visible state, I'm open to suggestions