ensure the animation plays out on warm launches
I observed that on warm starts the animation never gets to fully play out, it seems like it's dismissed before being able to finish which makes it look like something is broken.

Therefore we can register an exit animation listener and use the length of the animation to determine when we should actually move on which results in the animation being played to completion on warm starts.