Fix for the backwards typing bug (#5631)
Task/Issue URL: https://app.asana.com/0/488551667048375/1209104073191021

### Description

This PR fixes the backwards typing bug (typing from right-to-left) by
replacing the `shouldMoveCaretToStart` ViewState parameter with a
command, so that it's executed only once when the omnibar loses focus.

I've also removed `shouldMoveCaretToEnd` because it was not used
(enabled) anywhere in the code.

### Steps to test this PR

_The fix_
- [x] Start with a single NTP
- [x] Type something and verify it's left-to-right
- [x] Delete the text
- [x] Pull down the notification bar and change the dark mode setting
- [x] Pull up the notification bar
- [x] Type something in the address bar
- [x] Notice the text is correctly written left-to-right

_Long URL shown from the start_
- [x] Navigate to some website where the URL is longer than the address
bar
- [x] Tap on the webview so that the address bar loses focus
- [x] Verify that the URL is shown from the beginning, not just its end