Refactor: Update Tab Switcher to use TabSwitcherItem
The Tab Switcher now utilizes a new `TabSwitcherItem` data class to manage tabs.
The `TabSwitcherViewModel`'s `tabs` LiveData has been replaced with `tabSwitcherItems`, which returns a list of `TabSwitcherItem`.
The `render()` method in `TabSwitcherActivity` now receives a list of `TabSwitcherItem` instead of `TabEntity`.
The `updateTabGridItemDecorator()` now takes a tabSwitcherItemId.
The logic for detecting empty tabs in `onNewTabRequested()` has been updated to use `tabSwitcherItems` and `TabSwitcherItem.Tab`.
The `onCloseAllTabsConfirmed()` now iterates over `tabSwitcherItems` and checks if they are instances of `TabSwitcherItem.Tab`.
The `onTabDeleted` is updated to take TabSwitcherItem and delete the appropriate TabEntity
The menu item `closeAllTabs` visibility is now driven from `tabSwitcherItems`
The adapter now uses `TabSwitcherItem` instead of `TabEntity`