Refactor TabSwitcherAdapter to use lambda for binding adapter position
The `bindingAdapterPosition` is now passed as a lambda to `attachTabClickListeners`.

The code was changed to pass in the bindingAdapterPosition to the attachTabClickListeners.

The problem is that upon tab deletion we would now potentially hold an old bindingAdapterPosition. The fix is to evaluate the bindingAdapterPosition on every close click to ensure we have the most up to date position