Revert "SplashScreen Animation: Add wink animation (#5565)" removal (#5688)
Task URL: https://app.asana.com/0/488551667048375/1209483504518112

### Description

We removed the splashscreen animation while investigating issues due to
a user report of a black screen when launching. This reverts that
removal and will be a feature branch for adding upcoming mitigations to
reduce chances of this happening.

Task/Issue URL:
https://app.asana.com/0/1207908166761516/1209495201524764

### Description
Added a fallback mechanism to handle cases where the splash screen exit
animation fails to trigger. The fallback automatically proceeds to the
next screen after a 1.5-second delay if the normal exit animation
doesn't occur.

We now also only target Android 13 and above with the Splashscreen
animation after seeing mention of issues on Android 12L.

### Steps to test this PR

_Android 12L testing fallback_
- [ ] We can actually rely on the Android Studio launch bug to test this
- [ ] Create an emulator with API 32 (get one with the play store)
- [ ] Run the branch from the Android Studio
- [ ] The app should *not* show the splashscreen (remember we're using
the bug
- [ ] Check that `m_splashscreen_failed_to_launch` pixel fired
- [ ] After ~1.5 seconds the app should launch into onboarding/browser

_Android 12L testing normal launch_
- [ ] Close the app from before
- [ ] Locate the app on the home screen/app launcher
- [ ] Launch the app
- [ ] The static logo splashscreen should show
- [ ] The app should launch into onboarding/browser

_Test other OS versions_

Test that the splashscreen still works on other Android OS versions. A
suggested range could be 26, 30, 33, 35

On 33+
- [x] Open the app
- [x] Animated splashscreen should play
- [x] The app should launch into onboarding/browser

On <= 32
- [ ] Open the app
- [x] Static splashscreen logo should be visible
- [ ] The app should launch into onboarding/browser

### UI changes
No visual changes to the splash screen appearance or animation.