Fix alpha resetting for the TrackerAnimationTile
The previous commit fixed issues with drag and drop of tabs and glitches. Unfortunately it meant that the alpha state of the TrackerAnimationTile seemed to change as well, this seems to be due to the TabSwitcherItemDiffCallback being called and then triggering changes.

Now we pass through the isDragging state to the diff callback so that we can ensure that if we update the list that we maintain the TrackerAnimationTile alpha state, otherwise it would just reset to default.