Prevent setting animation tile seen if the first item is already a TrackerAnimationTile
This prevents a race condition when close all tabs is called and stops the animated tile staying visible due to the reactive changes that stem from setting the tab as seen/unseen.

We need this as a result of recent behaviour changes to close all tabs/fire button scenarios where we do NOT want the animated tile permanently dismissed. Only if a user presses the close button.