Multi-selection: Add new FAB to the tab switcher (#5467)
Task/Issue URL: https://app.asana.com/0/72649045549333/1207969573539951

### Description

This PR adds the FAB to the tab switcher. It is hidden under a feature
flag, which must be enabled first.

This PR demonstrates the 2 states of the FAB that will be used later in
the multi-selection -- "New Tab" & "Close Tabs". Clicking the button
will not actually change this in the actual implementation.

### Steps to test this PR

- [x] Go to Settings -> Feature Flag Inventory
- [x] Type "multi"
- [x] Enable the feature flag
- [x] Leave the tab switcher activity if there already
- [x] Open the tab switcher activity
- [x] Notice the FAB is displayed
- [x] Verify the `+` icon is shown and the title is "New Tab"
- [x] Scroll up & down
- [x] Verify that the FAB shrink while scrolling down
- [x] Tap on the FAB
- [x] Verify the icon changes to `x` and title says "Close Tabs"