Refresh sync settings screen automatically (#5773)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/608920331025315/task/1209669301264637

### Description
Allows the sync settings screen to automatically refresh while it is
being viewed. This is in part to provide a better UX to
https://github.com/duckduckgo/Android/pull/5762.

### Steps to test this PR

- [x] Enter sync settings with sync disabled. Verify everything looks as
expected and there is no visual updates to the UI
- [x] Enable sync. Verify everything looks as expected and there is no
visual updates to the UI.
- [x] Sync with another device, such that you can see two devices in
this view; your device, plus one other
- [x] On the other device, turn off sync
- [x] Back on the first device, verify it updates within the next 5s to
show the other device is now gone
- [x] Disable `sync / automaticallyUpdateSyncSettings` feature flag and
return to the sync settings screen. Repeat the above test and verify the
other device doesn’t automatically update since background refreshing is
disabled (i.e., fallback to original behaviour when flag disabled)
- [x] Smoke test any other related sync scenarios you think might be
problematic because of this

Co-authored-by: Craig Russell <1336281+CDRussell@users.noreply.github.com>