Refactor: Remove TabSwitcherTileAnimationMonitor
The `TabSwitcherTileAnimationMonitor` and its associated tests have been removed.
The logic to observe the animation tile visibility has been moved to the `TabSwitcherDataStore` to improve code organization.
The `TabSwitcherViewModel` now directly interacts with the `TabSwitcherDataStore`.
Also, the animation tile seen state is removed from `TabSwitcherDataStore`, the tile will be shown until is dismissed.
When clearing personal data, the animation tile will not be reset anymore.
Finally, the action to reset the animation tile has been updated in the dev settings.