Switch to using InfoPanel style for the tracker animation
It's the same view in both grid and list so we don't need two variations anymore.

I'm using a custom layout as adding the animation would have required a lot of hacky changes to the original InfoPanel. I'd rather look into that more in future if the experiment succeeds.

Renaming the item in the next commit