Throttle TabSwitcher UI updates
The `tabSwitcherItems` LiveData was causing excessive UI updates as on launch it would emit twice quickly due to the us updating the tabPreviewImage in the db from the tab you opened the TabSwitcher from.

This caused issues scrolling to the current active tab when launching the TabSwitcher.

This change introduces a `debounce` to the flow, conflating emissions within the debounce window and emit only the latest value, therefore reducing the update rate to the UI.