Tab Switcher Animation: Add placeholder animation (#5715)
Task/Issue URL:
https://app.asana.com/0/1207908166761516/1209368500700764

### Description

Adds a new tracker animation tile to the tab switcher behind a feature
flag that displays the number of trackers blocked in the last 7 days.
The tile appears at the top of the tab list and includes an animated
shield icon with a counter that animates up to show the total number of
trackers blocked.

The tile is permanent, a following PR will introduce dismissal with
show/hide logic.

The tracker count is currently a random number. A future PR will add the
real tracker count.

### Steps to test this PR

_Tab Switcher Animation_
- [x] Open tab switcher and verify tracker animation tile appears at top
- [x] Confirm tracker count animates up from partial to full count
- [x] Verify tile shows in both grid and list views
- [x] Check that tile cannot be dragged/reordered
- [x] Confirm tile alpha reduces when dragging other tabs
- [x] Verify shield animation plays correctly

_Tab Management_
- [x] Confirm tabs can still be reordered normally
- [x] Verify tab dragging and dropping works as expected
- [x] Check that tab deletion functions properly

[Designs

](https://www.figma.com/design/SngBHcIJci2HusCWGJuz0a/O-10%3A-Tab-Switcher?node-id=501-4212&m=dev)

### Demo

Note: Colours do not change for light and dark mode 

Phone


https://github.com/user-attachments/assets/cf37a27a-8bdc-453c-8528-5d502eb2798d

Tablet


https://github.com/user-attachments/assets/ab7ebd29-f611-4128-afa9-fd26dffe6d99

---------

Co-authored-by: Dax The Translator <daxmobile@duckduckgo.com>