always complete omnibar/minibar transitions with an animation (#5848)
Task/Issue URL: https://app.asana.com/0/0/1209754552084246

### Description
Adds logic to always complete the omnibar/minibar transition.

The behavior is as follows:
- If user scrolls down, we always immediately start transitioning to
minibar.
- If user scrolls up, we only start transitioning back to omnibar after
a scroll threshold since the start of last gesture is reached. This is
done so that as users scroll back-and-forth around the page we don't
switch between omnibar/minibar all the time - only significant scroll
gestures and flings, hitting top of the page, or clicking the minibar,
will reveal the omnibar back.
- If gesture has any velocity, we complete the transition in the
direction of the gesture.

Known issues:
- If we're near the bottom of the page, and **omnibar** is visible,
scrolling very slowly until hitting the bottom of the page might cause a
little bit of a bounce while we're force-transitioning to the minibar.