Experiment framework: Add minSdkVersion to Target  (#5850)
Task/Issue URL:
https://app.asana.com/0/1200581511062568/1209869840246080/f

### Description

Added support for minimum SDK version targeting in experiments. This
allows features to be enabled only for devices running a specific
Android SDK version or higher.

### Steps to test this PR

Added unit and integration tests 

_Test minSdkVersion target_
- [ ] install from this branch
- [ ] create a test feature
- [ ] create a remote config for the test feature that contains
`minSdkVersion` as target
- [ ] in a client that is on a SDK version lower than `minSdkVersion`
verify the feature is always disabled
- [ ] in a client that is on a SDK version equal or higher than
`minSdkVersion` verify the feature is enabled when the feature is
enabled in remote config