Fix omnibar touch not selecting text and showing keyboard
We moved to a use a container like view and wrapped content on the omnibarTextInput which meant that there was no click handling in when touching the text entry.

Previously we had omnibarTextInput match the full width and that would handle any touches, so we just do the same now and make it full width of our new container view.