Fix: Undo delete snackbar in the browser (#5908)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1202552961248957/task/1209975829243855?focus=true

### Description

This PR fixes the mechanism to display the undo delete snackbar in the
browser screen after all tabs are deleted in the tab switcher, when the
snackbar would not show up correctly.

The PR also fixes the position of the snackbar so that it's shown above
the omnibar if it's set to the bottom position.

### Steps to test this PR

- [x] Have 3 tabs and Omnibar set to bottom of the screen
- [x] Close a tab that is not selected
- [x] Switch to the last tab that is not selected while the snackbar is
showing
- [x] Verify no snackbar is displayed in the browser screen
- [x] Go back to the tab switcher
- [x] Close one of the tabs, then immediately close the last tab
- [x] Verify the undo snackbar is displayed for the last tab closed
- [x] Verify the snackbar is displayed above the bottom address bar

### UI changes

<image
src="https://github.com/user-attachments/assets/7816d82f-004e-4534-ba01-8c91e386d13b"
width="300" />