[WIP] Use sense of protection experiment to determine whether to show trackers blocked animation
I deleted AppPersonalityFeature which was the local toggle to determine whether to show the animation or not. I did not want to leave it in to cause any confusion with the experiment. It has not been completely removed in this PR which is why it's a WIP.

It would be good to confirm with Ana that the checks to show based on the new variants are correct as things have changed so much but I believe I have the checks correct.