update find in page feature in FadeOmnibarLayout (#5927)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1209874980847232

### Description

The "find in page" state is currently fully managed by
`BrowserTabViewModel`. This PR makes the omnibar aware of the visibility
portion of that state - just enough for it to react and pass it down to
its children.

Functionally, there’s no change for the production omnibar. But in the
experimental omnibar - where the "find in page" UI is moved inside the
card - we can now detect when it’s shown and hide other content
accordingly. This lets us reuse the omnibar card’s existing design and
animations.

This PR keeps the changes minimal, focused on adapting "find in page"
for the experimental omnibar. As a follow up, in
https://github.com/duckduckgo/Android/pull/5926, I'm attempting to
migrate the whole of "find in page" state management out of
`BrowserTabViewModel` and into the omnibar, along with further cleanup.