Add support for scam protection (#5853)
Task/Issue URL: https://app.asana.com/0/0/1209879813659870 

### Description

### Steps to test this PR

_Feature 1_
- [ ] Fresh install the app
- [ ] Using the app inspector, check that only phishing and malware
datasets are present. Data needs to be downloaded, so it might take a
few seconds to complete. The easiest way is to check the `revisions`
table in `malicious_sites` db
- [ ] Open
https://privacy-test-pages.site/security/badware/phishing.html and check
it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/malware.html
and check it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's not blocked

_Feature 2_
- [ ] Set `@Toggle.DefaultValue(true)` for `scamProtection`
- [ ] Fresh install the app
- [ ] Using the app inspector, check that phishing, malware and scam
datasets are present. Data needs to be downloaded, so it might take a
few seconds to complete. The easiest way is to check the `revisions`
table in `malicious_sites` db
- [ ] Open
https://privacy-test-pages.site/security/badware/phishing.html and check
it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/malware.html
and check it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's blocked
- [ ] Check `m_malicious-site-protection_error-page-shown` is fired with
`category=scam`
- [ ] Accept risk and visit site
- [ ] Check `m_malicious-site-protection_visit-site` is fired with
`category=scam`
- [ ] Click the omnibar icon
- [ ] Check privacy dashboard is shown normally
- [ ] Disable `scamProtection` under Feature Flag Inventory and restart
the app
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's not blocked

From #5857 

_Feature 1_
- [ ] Fresh install the app
- [ ] Using the app inspector, check that only phishing and malware
datasets are present. Data needs to be downloaded, so it might take a
few seconds to complete. The easiest way is to check the `revisions`
table in `malicious_sites` db
- [ ] Open
https://privacy-test-pages.site/security/badware/phishing.html and check
it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/malware.html
and check it's blocked
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's not blocked

_Feature 2_
- [ ] Without reinstalling the app, enable `scamProtection` under
Feature Flag Inventory and restart the app
- [ ] Using the app inspector, check that phishing, malware and scam
datasets are present. Data needs to be downloaded, so it might take a
few seconds to complete. The easiest way is to check the `revisions`
table in `malicious_sites` db
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's blocked.
- [ ] Disable `scamProtection` under Feature Flag Inventory and restart
the app
- [ ] Open https://privacy-test-pages.site/security/badware/scam.html
and check it's not blocked

### UI changes

![image](https://github.com/user-attachments/assets/d842d2c7-8b19-4329-885e-fcba2e4db8d0)



---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209910424512874

---------

Co-authored-by: Dax The Translator <daxmobile@duckduckgo.com>