Tile experiment text edge case (#5999)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1210140687033716

### Description
This PR fixes and edge case where only 1 tracker has been blocked and we
display the tile animation.
This only applies to English, other languages in this edge case will see
the english text.

Instructions -
[instructions.mp4](https://app.asana.com/app/asana/-/get_asset?asset_id=1210103609702441)
Install apk and open Settings / Experimental UI Settings
Enable any of the O-10 Experiments. Restart the app to ensure it works
as expected.
Note: This is not compatible with the O-A Experiment, please disable it
for proper testing

### Steps to test this PR

_1 tracker blocked_
- [ ] Fresh install and enable the Variant 2 in the O-10 experiment
- [ ] Visit a site that blocks 1 tracker
- [ ] Open the Tab Manager screen
- [ ] Tile should read “1 tracker blocked in the last 7 days"

### UI changes
| Before  | After |
| ------ | ----- |

![Screenshot_20250502_143738](https://github.com/user-attachments/assets/ff030e9a-adcc-4643-ba16-ba6098ad6ddf)|![Screenshot_20250502_144324](https://github.com/user-attachments/assets/0c3349f8-75b5-4de6-882c-1d5cb461bd40)|