Fix: Ensure the scrollbars are always restored (#6004)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207418217763355/task/1210108991225075?focus=true

### Description

This fixes a bug when the WebView scrollbars would stay hidden if the
preview generation threw an exception. This may happen fairly often with
tab swiping when you swipe between tabs fast so that the webview doesn't
get a chance to get laid out.

### Steps to test this PR

- [ ] Swipe tabs until you notice `Failed to generate WebView preview`
in the logcat
- [ ] Swipe back to the last tab
- [ ] Notice the WebView has scrollbars when you scroll it