Fix: NTP creation crash with tab swiping (#6007)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207418217763355/task/1210095537840719?focus=true

### Description

This PR fixes the NTP creation. There was a logical error in the
previous implementation when the wrong boolean was being returned in
`transformWhile()` when requesting a new tab, which in some situation
might result in the flow completion without emitting any value, which
resulted in a crash.

I added code that throws an exception when a new tab is not created
because it's an invalid state in the DB that I'm not sure how we should
recover from. I don't think that's what was happening but in case it
does, we'll know where to start looking.

### Steps to test this PR

I wasn't able to reproduce it, however, I added unit tests that should
verify the fix works as expected.