Fix RMF promo_single_action layout when vertical space is constrained (#6044)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210194248346098

### Description
The RMF `promo_single_action` was squished and misaligned in horizontal
orientation because it lacked the vertical space to fit all of its
elements in correct sizes.

To resolve that issue, we primarily need to make the new tab content
page scrollable, which coincidentally is already happening in
https://github.com/duckduckgo/Android/pull/6035, so I'm basing this PR
on top it.

Additionally:
1. I'm making sure that the new tab page uses the correct behaviors to
ensure that content is not clipped by the bottom navigation bar (when
the UI experiment is enabled).
 2. I'm updating the action button to `DaxButtonSecondary`.