Tab swiping: Improve state handling in the pager adapter (#6057)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1207418217763355/task/1210226097518220?focus=true

### Description

This PR updates the `TabPagerAdapter` to keep all data required for tab
fragment creation within the adapter and avoid DB lookups. Whenever tabs
are updated in the DB, the adapter data is updated.

It requires to keep some extra data besides the tab ID but eliminates
the chance for a race condition when the adapter requests a new fragment
and a tab is deleted before it has a chance get data.

### Steps to test this PR

Smoke testing tab swiping should be enough, as the race condition is not
easy to reproduce.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1208271988740971