fix an issue where nav bar would stay visible together with the keyboard when NTP was opened
When opening the new tab page, we initial don't show the keyboard. This scheduled a delayed task to execute and show the navigation bar. Before the time elapses to run the task, we automatically show the keyboard and focus on the omnibar, which hides the nav bar. However, the initially scheduled task would eventually execute and re-show the navigation bar even though the keyboard was also already visible.

The fix is to cancel the delayed task to show the navigation bar whenever the keyboard state changes.