Expand mutex setup to revisions and exceptions (#6049)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1202552961248957/task/1210226116146307?focus=true

### Description
We still see some ANRs that seem to be related to a database connection
not being acquired on first app run. Looking at the logs, it looks like
it happens on devices with 1GB RAM, which only allow for 1 simultaneous
SQLite connection.

Add mutex to get revisions, so if another DB operation needs to happen
at the same time, it can decide whether to wait for the lock to be
acquired or use a default value

### Steps to test this PR

_Feature 1_
- [ ] Smoke test malicious site protections