Add magnifying glass animation
A key point to this is the viewsExemptedFromOffset in TopAppBarBehavior. Without this we get a 56dp margin on the bottom and was a very frustrating thing to find!

Easier to wrap with constraintlayout and use a percentage to match designs.

The delay gives the keyboard a chance to pop, I tried using window inset listeners but I ended up with a white bar and could not figure out a way around it. Might be worth investigating more if we adopt this experiment.