fix an issue where omnibar would be stuck in wrong width after orientation change (#6109)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1157893581871903/task/1210317910216891?focus=true

### Description
Fixes an issue where omnibar would be stuck in wrong width after
orientation change.

Initially, I thought this is because of the state update which occurs
right after `onSizeChanged` and re-locks the dimensions but after
refactoring the code to ensure animation is started only when needed, I
realized that the issue was due to some race condition within the
Android framework while the layout params were being updated directly in
the `onSizeChanged` loop.

With the layout inspector, this is what I found out:
1. We are in an unfocused state, the card's content is `match_parent`,
as expected:
![Screenshot from 2025-05-21
11-36-51](https://github.com/user-attachments/assets/a8a463bb-b6f7-433b-b616-7f1dedf61a1f)
2. We click into the omnibar and grab focus, the card's contents get
locked so that they don't move while the card animates:
![Screenshot from 2025-05-21
11-37-05](https://github.com/user-attachments/assets/2f4d1685-7ac8-447a-833b-c98578cb86cb)
3. We rotate the screen and the dimensions get unlocked (they are
`match_parent` now) but they view doesn't adjust to the new size.
![Screenshot from 2025-05-21
11-37-15](https://github.com/user-attachments/assets/99c1d55c-b794-4b13-9755-94bbcc606b2b)
4. Only after requesting another layout change by dropping focus (or by
typing in another letter, or interacting with the omnibar in any other
way) the view adjusts to the new width:
![Screenshot from 2025-05-21
11-37-26](https://github.com/user-attachments/assets/9ff1a126-b714-4422-b60f-b1c1a77d6884)

I found it reliable to wait for the first render loop to finish before
unlocking the view, without having to refactor any animation state.

### Steps to test this PR

- [ ] Enable visual design flag.
- [ ] Focus on the omnibar and rotate the screen.
- [ ] Verify that the omnibar adjusts to the new width without needing
to drop focus.
- [ ] Open "find in page" and rotate the screen.
- [ ] Verify that the "find in page" layout adjusts to the new width
without needing to drop focus.
- [ ] When focus is dropped, the address bar shows without content
jumping significantly.