update the TabDataRepository to use a single source of truth for tab properties (#6254)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210530166473085?focus=true

### Description

Updates the `TabDataRepository` to use a single source of truth for tab
properties instead of creating independent database queries and
observable objects for both `LiveData` and `Flow` variants.

This resolves the original issue where `adClickManager` was receiving
new tabs without source URLs when tabs swiping feature was enabled:

https://github.com/duckduckgo/Android/blob/ef6a87d670c66f6764d8f97bd11b8ccc0cf03076/app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt#L1657-L1661

This issue was caused by `TabDataRepository#liveTabs` returning `null`
because with tabs swiping feature enabled there was no entity actually
observing the `LiveData` object so it was producing no values.

Now, even if only the `Flow` object is observed, it will transitively
activate the `LiveData` object and the `adClickManager` will receive the
correct tab properties.

This make the `LiveData` object the source of truth. I explored making
the `Flow` the source of truth but it would require a more complex
changes to the eager initialization of the `LiveData` that seem
unnecessary. We should remove the `LiveData` object in the future
whenever tab swiping feature flag is removed and it becomes the only
code path.

### Steps to test this PR

1. Hardcode a config with tab swiping enabled
(https://www.jsonblob.com/api/1384585817120497664).
2. Load https://www.search-company.site/#ad-id-5.
3. Click on the ad.
4. Click on the "Publisher site".
5. Click on the "Green T-shirt". This will open in a new tab.
6. Verify that `convert.ad-company.site` requests are loaded.


https://github.com/user-attachments/assets/7ea2f9eb-a713-47d4-a500-54961ff9c763

---------

Co-authored-by: 0nko <oruttkay@duckduckgo.com>