add input mode switch user preference to Duck.ai settings (internal only) (#6291)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210636235167313?focus=true

### Description
Separates the new Duck.ai Input Screen feature from the Visual Design
data store and flags, and moves it from the experimental options to the
main body of the app.

- Removes the `experimentalUITheming#duckAIPoCFeature` feature flag and
creates a new `aiChat#duckAiInputScreen` flag instead.
- Enables the new flag for all internal users.
- Adds a new "Input Screen" user preference to Duck.ai settings screen -
this settings option is only be available if the feature flag is enabled
(so only for internal users).
- When the Input Screen preference is toggled on, it also enables the
visual design, as there's a dependency between the visual design and the
new Input Screen.

### Steps to test this PR

- [ ] Clean install the app.
- [ ] Verify that clicking on the omnibar just focuses it.
- [ ] Settings -> Duck.ai -> enable Search input.
- [ ] Verify that clicking on the omnibar opens the new Input Screen
with mode switch.
- [ ] Settings -> Experimental UI Settings -> disable Visual Design
updates.
- [ ] Verify that Settings -> Duck.ai -> Search Input is disabled and
that clicking on the omnibar doesn't open the Input Screen.
- [ ] Change build variant to `play`.
- [ ] Verify that the Input Screen is not available in the Duck.ai
settings screen.