Fix logo visibility on new tab page based on onboarding status (#6326)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1202552961248957/task/1210683053177465?focus=true

### Description

This PR modifies the logic for displaying the home background logo in
the new tab view. Previously, the logo was only shown when there were no
messages and no favorites. Now, it will also be shown during onboarding,
regardless of message status, as long as there are no favorites.

### Steps to test this PR

_Logo Visibility_
- [x] Verify the home background logo appears during onboarding even
when messages are present
- [x] Verify the logo still appears when there are no favorites and no
messages
- [x] Verify the logo is hidden when favorites are present

### UI changes
N/A