Fix ANR/Crash on autoconsent init handler (#6354)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/488551667048375/task/1210704553451145?focus=true

### Description
This PR tries to fix and ANR and crash in the autoconsent init handler.
It does remove the call to trimIndent() which is not very performant for
long strings (and doesn't seem to be needed) and it moves all the logic
to IO leaving only the webview call to evaluate JS in main.

### Steps to test this PR

- [ ] Test autoconsent