add OS version check for Input Screen overshoot alpha transitions (#6404)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210786203196179?focus=true

### Description
Only uses the overshoot interpolator on newer Android version to avoid
an issue where alpha channel is zero (or generally cause a black screen
for a split second) when the interpolation goes over the <0.0, 1.0>
bounds.

I'm also noticing that translation transition is not available at all on
Android versions < 31, at least based on emulator testing, but leaving
it in doesn't seem to cause any other issue.

### Steps to test this PR

- [x] Try the Input Screen transitions on a device with API >= 34
- [x] Try the Input Screen transitions on a device with API < 34