Fix ignored files committed
Spotless breaks because we have classes and files that have been commited that should technically be ignored by our .gitignore settings. See https://github.com/diffplug/spotless/issues/911#issuecomment-904887074

Therefore we have to either remove these files or move them.

In this case the DeviceShield stuff was in a package called "report" which we ignore in .gitignore. Therefore I gave it a new package name

We want to keep the icon and I guess we want to keep the copyright folder so I made that work in .gitignore