compute Input Screen tab indicator height to always be vertically centered in the tab switcher (#6435)
Task/Issue URL:
https://app.asana.com/1/137249556945/project/1208671518894266/task/1210767017822401

### Description
Previously, the dimensions picked for the tab switcher and its tab
indicators often resulted in the switcher having a height of an odd
number of pixels while the indicator had height of even number of
pixels. This resulted in the pill not being perfectly vertically aligned
within the switcher. It depends on the pixel density of the device, but
happened often on the devices we tested on.

This PR adds a bit of logic to subtract a pixel when needed, to allow
for that perfect vertical alignment.

### Steps to test this PR

- [ ] Try a commit before this change and verify that the space below
the tab indicator is bigger than the space above it.
- [ ] Try this change and verify that the space is now equal.